fix #94 Boost on attribute can't be a decimal, DOH!

+++ src/Nest/Domain/Mapping/ElasticPropertyAttribute.cs
@@ -19,7 +19,7 @@ namespace Nest
 		public TermVectorOption TermVector { get; set; }
 		public FieldIndexOption Index { get; set; }
 
-		public decimal Boost { get; set; }
+		public double Boost { get; set; }
 
 		public string Analyzer { get; set; }
 		public string IndexAnalyzer { get; set; }
